Changed penrose.sh to allow running in foreground.
git-svn-id: http://unused-160-33.aus.redhat.com/svn/penrose-server/trunk@323 04c32f3d-caf7-0310-9fe0-d49f5cee804d

+++ bin/penrose.sh
@@ -152,22 +152,24 @@ cd "$PENROSE_HOME"
 mkdir -p "$PENROSE_HOME/var"
 PENROSE_PID="$PENROSE_HOME/var/penrose.pid"
 
-start() {
+if [ "$1" = "start" ] ; then
+
   if [ -f "$PENROSE_PID" ] ; then
     echo Penrose Server is running
     exit 1
   else
-    if [ -n "$CYGHOME" ] ; then
-      exec "$JAVACMD" $PENROSE_DEBUG_OPTS $PENROSE_OPTS -classpath "$LOCALCLASSPATH" -Djavax.management.builder.initial=mx4j.server.MX4JMBeanServerBuilder -Dpenrose.home="$PENROSE_HOME" -Dcygwin.user.home="$CYGHOME" org.safehaus.penrose.PenroseServer $PENROSE_ARGS "$@" >> "$PENROSE_HOME/var/penrose.out" 2>&1 &
-    else
-      exec "$JAVACMD" $PENROSE_DEBUG_OPTS $PENROSE_OPTS -classpath "$LOCALCLASSPATH" -Djavax.management.builder.initial=mx4j.server.MX4JMBeanServerBuilder -Dpenrose.home="$PENROSE_HOME" org.safehaus.penrose.PenroseServer $PENROSE_ARGS "$@" >> "$PENROSE_HOME/var/penrose.out" 2>&1 &
-    fi
+    shift
+    exec "$JAVACMD" $PENROSE_DEBUG_OPTS $PENROSE_OPTS \
+    -classpath "$LOCALCLASSPATH" \
+    -Dpenrose.home="$PENROSE_HOME" \
+    org.safehaus.penrose.PenroseServer $PENROSE_ARGS "$@" \
+    >> "$PENROSE_HOME/var/penrose.out" 2>&1 &
 
     echo $! > "$PENROSE_PID"
   fi
-}
 
-stop() {
+elif [ "$1" = "stop" ] ; then
+
   if [ -f "$PENROSE_PID" ] ; then
     kill -9 `cat "$PENROSE_PID"` > /dev/null 2>&1
     rm -f "$PENROSE_PID"
@@ -175,47 +177,13 @@ stop() {
     echo Penrose Server is not running
     exit 1
   fi
-}
-
-status() {
-  if [ -f "$PENROSE_PID" ] ; then
-    echo Penrose Server is running
-  else
-    echo Penrose Server is not running
-  fi
-}
-
-usage() {
-  echo "Usage: penrose.sh COMMAND"
-  echo
-  echo "Commands:"
-  echo "  start             Start Penrose Server"
-  echo "  stop              Stop Penrose Server"
-  echo "  restart           Restart Penrose Server"
-  echo "  status            Check Penrose Server status"
-  exit 1
-}
-
-if [ "$1" = "start" ] ; then
-
-  start
-
-elif [ "$1" = "stop" ] ; then
-
-  stop
-
-elif [ "$1" = "restart" ] ; then
-
-  stop
-  start
-
-elif [ "$1" = "status" ] ; then
-
-  status
 
 else
 
-  usage
+  exec "$JAVACMD" $PENROSE_DEBUG_OPTS $PENROSE_OPTS \
+  -classpath "$LOCALCLASSPATH" \
+  -Dpenrose.home="$PENROSE_HOME" \
+  org.safehaus.penrose.PenroseServer $PENROSE_ARGS "$@"
 
 fi
